Moved updateSaveButtons to chat-message.js + converted indentation to fit the rest of the project. Sorry, my editor defaults to spaces. I think we really need eslint to enforce the indentation wanted at some point.
